fix(api/etl): enhance ETL to process CSVs in chunks
